% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_posteriori.R
\name{a_posteriori}
\alias{a_posteriori}
\title{Vector of a posteriori density values of ability level}
\usage{
a_posteriori(apriori, prob)
}
\arguments{
\item{apriori}{a vector of a priori density function values of the evaluated
ability levels}

\item{prob}{a vector of probability response for every evaluated ability
level given the item response}
}
\value{
A vector of a posteriori density values
}
\description{
Computes the a posteriori density function values of the evaluated ability
levels given the item response
}
\author{
Javier Rodríguez-Cuadrado
}
