% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ability_density.R
\name{ability_density}
\alias{ability_density}
\title{Vector of density values of ability level}
\usage{
ability_density(dens, ...)
}
\arguments{
\item{dens}{density function (e.g. dnorm, dunif, etc.)}

\item{...}{parameters of the density function}
}
\value{
A vector of density values
}
\description{
Computes the density function values of the evaluated ability levels
}
\author{
Javier Rodríguez-Cuadrado
}
