% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_sons.R
\name{allocate_sons}
\alias{allocate_sons}
\title{Allocate sons in the CAT decision tree}
\usage{
allocate_sons(nodes_prev, nodes, level)
}
\arguments{
\item{nodes_prev}{list of node lists of the nodes from the previous level}

\item{nodes}{list of node lists of the nodes from the current level}

\item{level}{level of the CAT decision tree}
}
\value{
A list of node lists updated with the information of the sons
}
\description{
Fills the information of the sons of the previous level nodes
}
\author{
Javier Rodr?guez-Cuadrado
}
