% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_last_level.R
\name{create_last_level}
\alias{create_last_level}
\title{CAT decision tree last level generator}
\usage{
create_last_level(nodes_prev, nres, level, prob_array, SE)
}
\arguments{
\item{nodes_prev}{list of node lists of the nodes from the previous level}

\item{nres}{vector of number of possible responses for every item}

\item{level}{last-level number (equals the length of the test plus one)}

\item{prob_array}{3-D array of probability responses. Dim 1 represent items,
dim 2 represent evaluated ability levels and dim 3 represent possible
responses}

\item{SE}{minimum standard error of the ability level}
}
\value{
A list of lists. Each of these lists represent a node of the
last level of the decision tree
}
\description{
Generates a list of node lists for the last level of the CAT decision tree
}
\author{
Javier Rodríguez-Cuadrado
}
