% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher_GRM.R
\name{Fisher_GRM}
\alias{Fisher_GRM}
\title{Fisher Information under GRM}
\usage{
Fisher_GRM(theta_est, item_par, nres)
}
\arguments{
\item{theta_est}{ability level}

\item{item_par}{vector containing the item parameters. First component
is the \code{alpha} parameter and the next are the \code{beta} parameters}

\item{nres}{number of possible item responses}
}
\value{
An integer that represents the Fisher Information value of the
specified item given the ability level
}
\description{
Computes the item Fisher Information given an ability level based on the GRM
model
}
\author{
Javier Rodriguez-Cuadrado
}
