% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemBank.R
\docType{data}
\name{itemBank}
\alias{itemBank}
\title{Example item bank}
\format{
An object of class \code{data.frame}.
}
\usage{
data(itemBank)
}
\description{
Item bank data generated using \code{genPolyMatrix} from \code{catR} package.
}
\keyword{datasets}
