% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemRes.R
\docType{data}
\name{itemRes}
\alias{itemRes}
\title{Example item responses}
\format{
An object of class \code{matrix}.
}
\usage{
data(itemRes)
}
\description{
Item responses data to test with \code{data(itemBank)}.
There are 30 subjects and their responses to 100 items.
}
\keyword{datasets}
