% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{plot_tree}
\alias{plot_tree}
\title{CAT decision tree plot}
\usage{
plot_tree(object, levels = 3, tree = 1)
}
\arguments{
\item{object}{A \code{cat.dt} object}

\item{levels}{Number of levels to plot, starting from the first one.}

\item{tree}{Index of tree to plot. The total number of trees is given by \code{length(nodes$nodes[[1]])}.}
}
\value{
A ggplot2 object
}
\description{
Generates a plot object to visualize the CAT decision tree
}
\author{
Javier Rodríguez-Cuadrado
}
