% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probab_NRM.R
\name{probab_NRM}
\alias{probab_NRM}
\title{Item response NRM probabilities}
\usage{
probab_NRM(item_par, nres)
}
\arguments{
\item{item_par}{vector containing the item parameters. Odd components are
the \code{alpha} parameters and even are the \code{beta} parameters}

\item{nres}{number of possible item responses}
}
\value{
A matrix of response probabilities. Rows represent evaluated ability
levels and columns represent responses
}
\description{
Computes the probabilities of picking every possible response of an
specified item from the item bank for all evaluated ability levels
using the Nominal Response Model
}
\author{
Javier Rodríguez-Cuadrado
}
