% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAT_summary.R
\name{summary.cat.dt}
\alias{summary.cat.dt}
\title{CAT summary}
\usage{
\method{summary}{cat.dt}(object, ...)
}
\arguments{
\item{object}{A \code{cat.dt} object}

\item{...}{not used}
}
\value{
A summary of the \code{cat.dt} object
}
\description{
Summary of the \code{cat.dt} object generated
}
\author{
Javier Rodr?guez-Cuadrado
}
