% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.dt.R
\docType{package}
\name{cat.dt}
\alias{cat.dt}
\alias{cat.dt-package}
\title{cat.dt: A package implementing the Merged Tree-CAT method,
for computerized adaptive testing based on a tree-like structure.}
\description{
The cat.dt package provides a fast method such that very complex and deep
tests can be generated using the resources that are available in a
personal computer.
}
\section{Main interface function}{

CAT_DT
}

\seealso{
\code{\link{CAT_DT}}
}
\author{
Javier Rodríguez-Cuadrado, David Delgado-Gómez, Juan C. Laria
}
