% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Ability level estimation}
\usage{
estimate(dens_vec)
}
\arguments{
\item{dens_vec}{vector of density function values of the evaluated ability
levels}
}
\value{
A number, the expected value of the ability level density function
}
\description{
Computes the estimated ability level given the ability level density
function values
}
\author{
Javier Rodríguez-Cuadrado
}
